+++ src/test/java/org/opentripplanner/transit/raptor/_data/transit/TestTransferPoint.java
@@ -6,11 +6,19 @@ import org.opentripplanner.model.transfer.TransferPoint;
 public class TestTransferPoint implements TransferPoint {
 
   private final int stop;
+  // Wer need this because trip pattern can pass through same stop more than once
+  private final int stopPosition;
   private final TestTripSchedule schedule;
   private final boolean applyToAllTrips;
 
-  public TestTransferPoint(int stop, TestTripSchedule schedule, boolean applyToAllTrips) {
+  public TestTransferPoint(
+    int stop,
+    int stopPosition,
+    TestTripSchedule schedule,
+    boolean applyToAllTrips
+  ) {
     this.stop = stop;
+    this.stopPosition = stopPosition;
     this.schedule = schedule;
     this.applyToAllTrips = applyToAllTrips;
   }
@@ -26,11 +34,11 @@ public class TestTransferPoint implements TransferPoint {
   }
 
   public int getStopPosition() {
-    return schedule.pattern().findStopPositionAfter(0, stop);
+    return this.stopPosition;
   }
 
-  public boolean matches(TestTripSchedule schedule, int stop) {
-    return this.schedule == schedule && this.stop == stop;
+  public boolean matches(TestTripSchedule schedule, int stop, int stopPosition) {
+    return this.schedule == schedule && this.stop == stop && this.stopPosition == stopPosition;
   }
 
   @Override